+++ atari/src/include/fujinet-fuji-atari.h
@@ -1,0 +1,10 @@
+#ifndef FUJINET_FUJI_ATARI_H
+#define FUJINET_FUJI_ATARI_H
+
+#include <stdbool.h>
+#include <stdint.h>
+
+bool fuji_success();
+extern uint8_t *_t_fuji_open_appkey;
+
+#endif // FUJINET_FUJI_ATARI_H