+++ commodore/src/fn_fuji/fuji_appkey_open_read.c
@@ -1,0 +1,8 @@
+#include <stdbool.h>
+#include <stdint.h>
+#include "fujinet-fuji.h"
+
+bool fuji_read_appkey(uint16_t creator_id, uint8_t app_id, uint8_t key_id, uint16_t *count, uint8_t *data, enum AppKeySize keysize)
+{
+	return true;
+}