+++ commodore/src/fn_fuji/fuji_appkey_read.c
@@ -1,8 +1,0 @@
-#include <stdbool.h>
-#include <stdint.h>
-#include "fujinet-fuji.h"
-
-bool fuji_appkey_read(AppKeyRead *buffer)
-{
-	return 0;
-}