+++ commodore/src/fn_fuji/fuji_appkey_write.c
@@ -1,8 +1,0 @@
-#include <stdbool.h>
-#include <stdint.h>
-#include "fujinet-fuji.h"
-
-bool fuji_appkey_write(uint16_t count, AppKeyWrite *buffer)
-{
-	return 0;
-}