+++ common/src/network_init.c
@@ -5,7 +5,7 @@
 
 #ifdef __APPLE2__
 #include "fujinet-bus-apple2.h"
-#include "apple2/src/bus/inc/sp.h"
+#include "sp.h"
 #endif
 
 uint8_t network_init()